+++ allure-results/038b7455-5fb2-49fc-bd57-3a2ac3dd92e7-result.json
@@ -1,0 +1,1 @@
+{"uuid":"038b7455-5fb2-49fc-bd57-3a2ac3dd92e7","historyId":"fd74ddc9b455d1f2819f903df16fea00","fullName":"Sign in function: Attempt to login with invalid information","labels":[{"name":"tag","value":"AUT-4"},{"name":"tag","value":"InvalidSignIn"},{"name":"host","value":"Nessa"},{"name":"thread","value":"28028@Nessa.main(1)"},{"name":"feature","value":"Sign in function"},{"name":"story","value":"Attempt to login with invalid information"},{"name":"suite","value":"Sign in function"},{"name":"testClass","value":"Attempt to login with invalid information"},{"name":"framework","value":"cucumber4jvm"},{"name":"language","value":"java"},{"name":"gherkin_uri","value":"file:///D:/ten10/programming/GroupWork/src/test/resources/Features/SignIn.feature"},{"name":"package","value":"src.test.resources.Features.SignIn_feature.Sign in function"}],"links":[],"name":"Attempt to login with invalid information","status":"passed","stage":"finished","steps":[{"name":"Given  I have navigated to the login page","status":"passed","statusDetails":{"known":false,"muted":false,"flaky":false},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1671104449784,"stop":1671104451102},{"name":"When  I input invalid details","status":"passed","statusDetails":{"known":false,"muted":false,"flaky":false},"stage":"finished","steps":[],"attachments":[{"name":"Data table","source":"a8aa6529-7c37-45f6-8bb5-de8c13af3901-attachment.csv","type":"text/tab-separated-values"}],"parameters":[],"start":1671104451102,"stop":1671104451193},{"name":"And  Hit Enter","status":"passed","statusDetails":{"known":false,"muted":false,"flaky":false},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1671104451193,"stop":1671104451329},{"name":"Then  I am not logged in","status":"passed","statusDetails":{"known":false,"muted":false,"flaky":false},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1671104451329,"stop":1671104451528}],"attachments":[],"parameters":[],"start":1671104449784,"stop":1671104451528}
