+++ allure-results/1a89fc1e-7ff9-4472-bbf8-3ab60e027c03-result.json
@@ -1,0 +1,1 @@
+{"uuid":"1a89fc1e-7ff9-4472-bbf8-3ab60e027c03","historyId":"d9833b814ae57fc58edabd7ea353216b","fullName":"Sign in function: Logging in with valid details","labels":[{"name":"tag","value":"AUT-4"},{"name":"tag","value":"SignIn"},{"name":"host","value":"Nessa"},{"name":"thread","value":"28028@Nessa.main(1)"},{"name":"feature","value":"Sign in function"},{"name":"story","value":"Logging in with valid details"},{"name":"suite","value":"Sign in function"},{"name":"testClass","value":"Logging in with valid details"},{"name":"framework","value":"cucumber4jvm"},{"name":"language","value":"java"},{"name":"gherkin_uri","value":"file:///D:/ten10/programming/GroupWork/src/test/resources/Features/SignIn.feature"},{"name":"package","value":"src.test.resources.Features.SignIn_feature.Sign in function"}],"links":[],"name":"Logging in with valid details","status":"passed","stage":"finished","steps":[{"name":"Given  I have navigated to the login page","status":"passed","statusDetails":{"known":false,"muted":false,"flaky":false},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1671104446593,"stop":1671104449078},{"name":"When  I input valid details","status":"passed","statusDetails":{"known":false,"muted":false,"flaky":false},"stage":"finished","steps":[],"attachments":[{"name":"Data table","source":"7fc04099-2e77-4caf-ae8a-537e0aa688dd-attachment.csv","type":"text/tab-separated-values"}],"parameters":[],"start":1671104449079,"stop":1671104449187},{"name":"And  Hit Enter","status":"passed","statusDetails":{"known":false,"muted":false,"flaky":false},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1671104449187,"stop":1671104449525},{"name":"Then  I am logged in successfully","status":"passed","statusDetails":{"known":false,"muted":false,"flaky":false},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1671104449525,"stop":1671104449730}],"attachments":[],"parameters":[],"start":1671104446590,"stop":1671104449731}
