+++ allure-results/36ccba96-790d-45cd-aac4-5dd5b6721f0c-container.json
@@ -1,0 +1,1 @@
+{"uuid":"36ccba96-790d-45cd-aac4-5dd5b6721f0c","name":"Scenario: Attempt to login with invalid information","children":["038b7455-5fb2-49fc-bd57-3a2ac3dd92e7"],"befores":[],"afters":[],"start":1671104449784,"stop":1671104451528}
