+++ allure-results/7101d465-47a7-44c7-950c-e76e437964ee-container.json
@@ -1,0 +1,1 @@
+{"uuid":"7101d465-47a7-44c7-950c-e76e437964ee","name":"Scenario: Logging in with valid details","children":["1a89fc1e-7ff9-4472-bbf8-3ab60e027c03"],"befores":[],"afters":[],"start":1671104446590,"stop":1671104449732}
