+++ allure-results/86ba9733-7d0e-4ce4-91d1-03bcbaff0e66-result.json
@@ -1,0 +1,1 @@
+{"uuid":"86ba9733-7d0e-4ce4-91d1-03bcbaff0e66","historyId":"c0d24f2ffb9c41dcdc9e7bc83e2cf158","fullName":"Sign in function: Attempt to login with empty fields","labels":[{"name":"tag","value":"AUT-4"},{"name":"tag","value":"emptyFields"},{"name":"host","value":"Nessa"},{"name":"thread","value":"28028@Nessa.main(1)"},{"name":"feature","value":"Sign in function"},{"name":"story","value":"Attempt to login with empty fields"},{"name":"suite","value":"Sign in function"},{"name":"testClass","value":"Attempt to login with empty fields"},{"name":"framework","value":"cucumber4jvm"},{"name":"language","value":"java"},{"name":"gherkin_uri","value":"file:///D:/ten10/programming/GroupWork/src/test/resources/Features/SignIn.feature"},{"name":"package","value":"src.test.resources.Features.SignIn_feature.Sign in function"}],"links":[],"name":"Attempt to login with empty fields","status":"passed","stage":"finished","steps":[{"name":"Given  I have navigated to the login page","status":"passed","statusDetails":{"known":false,"muted":false,"flaky":false},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1671104451555,"stop":1671104452840},{"name":"When  Hit Enter","status":"passed","statusDetails":{"known":false,"muted":false,"flaky":false},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1671104452840,"stop":1671104452907},{"name":"Then  An error message appears","status":"passed","statusDetails":{"known":false,"muted":false,"flaky":false},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1671104452908,"stop":1671104453109}],"attachments":[],"parameters":[],"start":1671104451555,"stop":1671104453109}
