+++ allure-results/a9129089-6ff0-4ae1-858b-70a9bd4c249f-container.json
@@ -1,0 +1,1 @@
+{"uuid":"a9129089-6ff0-4ae1-858b-70a9bd4c249f","name":"Scenario: Attempt to login with empty fields","children":["86ba9733-7d0e-4ce4-91d1-03bcbaff0e66"],"befores":[],"afters":[],"start":1671104451555,"stop":1671104453109}
