+++ src/test/java/StepDefinitions/Hooks.java
@@ -1,0 +1,23 @@
+package StepDefinitions;
+
+import Util.DriverFactory;
+import io.cucumber.java.After;
+import io.cucumber.java.Before;
+
+public class Hooks {
+
+    public static DriverFactory driverFactory = new DriverFactory();
+
+    @Before
+    public void before(){
+        driverFactory.getDriver();
+    }
+
+    @After
+    public void after(){
+        driverFactory.quitDriver();
+    }
+
+
+
+}
